+++ rooms/views.py
@@ -1,9 +1,13 @@
-from django.shortcuts import render, redirect, HttpResponse
+from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from . forms import RoomForm
 from . models import Room
 from posts.models import Post 
 
 # Create your views here.
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -21,9 +25,14 @@ def room(request, pk):
     context = {'room': room, 'posts':posts,}
     return render(request, 'rooms/room.html', context)
 
+@login_required(login_url='login')
 def editRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('User access denied')
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -33,8 +42,11 @@ def editRoom(request, pk):
     context = {'form':form}
     return render(request, 'rooms/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('User access denied')
     if request.method == "POST":
         room.delete()
         return redirect('home')
